move items route from main
